refactor[w1.5]refactoring and implementing methods in mysql repo layers

+++ HealthcareSystem/Backend/Service/UsersAndWorkingTime/PatientService.cs
@@ -25,11 +25,12 @@ namespace Service.UsersAndWorkingTime
         }
         public Patient RegisterPatient(Patient patient)
         {
-            if (!patient.IsGuest &&  (!IsUsernameValid(patient.Username) || !IsPasswordValid(patient.Password)))
+            if (!patient.IsGuest && (!IsUsernameValid(patient.Username) || !IsPasswordValid(patient.Password)))
             {
                 throw new BadRequestException("Your username or password is incorrect. Please try again.");
             }
-            return _activePatientRepository.AddPatient(patient);
+            _activePatientRepository.AddPatient(patient);
+            return patient;
         }
 
         public Patient EditPatient(Patient patient)
@@ -38,19 +39,20 @@ namespace Service.UsersAndWorkingTime
             {
                 return null;
             }
-            return _activePatientRepository.SetPatient(patient);
+            _activePatientRepository.SetPatient(patient);
+            return patient;
         }
 
         public bool DeletePatient(string jmbg)
         {
-            Patient deletedPatient = _activePatientRepository.DeletePatient(jmbg);
-            if (deletedPatient != null)
+            Patient patient = _activePatientRepository.GetPatientByJmbg(jmbg);
+            if (patient == null)
             {
-                Patient newPatient = _deletedPatientRepository.AddPatient(deletedPatient);
-                if (newPatient != null)
-                    return true;
+                return false;
             }
-            return false;
+            _activePatientRepository.DeletePatient(jmbg);
+            _deletedPatientRepository.AddPatient(patient);
+            return true;
         }
 
         public List<Patient> ViewPatients()
